requirements.txt
numpy==1.20.1
requests==2.25.1
pymongo==3.11.3
redis==3.5.3
aiogram==2.8
pytesseract==0.3.7
opencv-python==4.5.1.48
python-dotenv==0.15.0
Pillow==8.1.2
Wand==0.6.6
PyYAML==5.4.1
